label to type long

+++ Code/Utilities/MUTr2_TrainModel_Sub.py
@@ -78,7 +78,7 @@ def GNNtrain(model, Sample,optimizer):
     for data in Sample:
         out = model(data.x, data.edge_index, data.edge_attr, data.batch)
         # data.y = torch.argmax(data.y, dim=1)
-        loss = criterion(out, data.y)
+        loss = criterion(out, data.y.long())
         loss.backward()  # Derive gradients.
         optimizer.step()  # Update parameters based on gradients.
         optimizer.zero_grad()
@@ -95,7 +95,7 @@ def GNNvalidate(model, Sample):
          y_index = data.y.argmax(dim=1)
          correct += int((pred == y_index).sum())  # Check against ground-truth labels.
          data.y = torch.argmax(data.y, dim=1)
-         loss = criterion(out, data.y)
+         loss = criterion(out, data.y.long())
          loss_accumulative += float(loss)
     return (correct / len(Sample.dataset), loss_accumulative/len(Sample.dataset))
